update whole project with : 1)mapper 2)proper Architecture 3)added Category etc....

+++ src/main/java/com/example/blog/service/PostService.java
@@ -1,11 +1,23 @@
 package com.example.blog.service;
 
 
+import com.example.blog.dto.PostRequest;
+import com.example.blog.dto.PostResponse;
+import com.example.blog.exception.ResourceNotFoundException;
+import com.example.blog.exception.UnauthorizedActionException;
+import com.example.blog.mapper.PostMapper;
 import com.example.blog.model.Post;
+import com.example.blog.model.PostStatus;
+import com.example.blog.model.User;
+import com.example.blog.repository.CommentRepository;
 import com.example.blog.repository.PostRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.config.ConfigDataResourceNotFoundException;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
+import java.security.InvalidParameterException;
 import java.util.List;
 import java.util.Optional;
 
@@ -14,32 +26,73 @@ import java.util.Optional;
 public class PostService {
 
     private final PostRepository postRepository;
+    private final PostMapper postMapper;
+    private final UserService userService;
+    private final S3Service s3Service;
+    private final CommentRepository commentRepository;
 
-    public Post savePost(Post post){
-        return postRepository.save(post);
+
+    //create post
+    public PostResponse savePost(PostRequest postRequest, String email) {
+        User user = userService.findByEmail(email)
+                .orElseThrow(() -> new RuntimeException("User not found"));
+        Post post = postMapper.toEntity(postRequest);
+        post.setUser(user);
+        Post savedPost = postRepository.save(post);
+        PostResponse postResponse = postMapper.toResponse(savedPost);
+        return postResponse;
     }
 
-    public Optional<Post> findById(Long id){
-        return postRepository.findById(id);
+    //get post by id
+    public PostResponse findById(Long id) {
+        Post post = postRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Post not Found for post Id" + id));
+        return postMapper.toResponse(post);
     }
 
-    public List<Post> findAllPosts(){
-        return postRepository.findAll();
+    //get all post
+    public List<PostResponse> findAllPosts() {
+        List<Post> posts = postRepository.findAll();
+        return posts.stream().map(postMapper::toResponse).toList();
     }
 
-    public List<Post> findByUserId(Long userId){
-        return postRepository.findByUserId(userId);
+    //get post by user id
+    public List<PostResponse> findByUserId(Long userId) {
+        List<Post> posts = postRepository.findByUserId(userId);
+        return posts.stream().map(postMapper::toResponse).toList();
     }
 
-    public List<Post> findPublishedPost(){
-        return postRepository.findByStatus("PUBLISHED");
+    //update post
+    public PostResponse updatePost(Long id, PostRequest postRequest, String email) {
+        Post post = postRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Post Not Found for the post id " + id));
+        if (!post.getUser().getEmail().equals(email)) {
+            throw new UnauthorizedActionException("You can only modify your own posts!");
+        }
+        postMapper.updateEntityFromDto(postRequest, post);
+        post.setStatus(PostStatus.valueOf(postRequest.getStatus()));
+
+        Post savedPost = postRepository.save(post);
+        return postMapper.toResponse(savedPost);
+
+
     }
 
-    public void deletePost(Long id){
+    public List<PostResponse> findPublishedPost() {
+        List<Post> publishedPost = postRepository.findByStatus(PostStatus.PUBLISHED);
+        return publishedPost.stream().map(postMapper::toResponse).toList();
+    }
+    @Transactional
+    public void deletePost(Long id, String email) {
+        Post post = postRepository.findById(id).orElseThrow(() ->
+                new ResourceNotFoundException("Post Not Found for the post id " + id));
+        if (!post.getUser().getEmail().equals(email)) {
+            throw new UnauthorizedActionException("You can only modify your own posts!");
+        }
+        if (post.getImageUrl() != null) {
+            s3Service.deleteFile(post.getImageUrl());
+        }
+        commentRepository.deleteByPostId(post.getId());
         postRepository.deleteById(id);
     }
-
-
 
 
 }
